Download then reinject latest mappings
