+++ main.py
@@ -1,4 +1,4 @@
-import os, asyncio, sys, time
+import os, asyncio, sys, time, aiohttp, json
 
 from watchdog.events import FileSystemEventHandler
 from watchdog.observers import Observer
@@ -24,6 +24,38 @@ except:
     pass
 
 Initialized = False
+
+SUCCESSFUL_FETCH_THIS_RUN = False
+
+async def fetch_class_mappings(css_translations_path : str, loader : Loader):
+    global SUCCESSFUL_FETCH_THIS_RUN
+
+    if SUCCESSFUL_FETCH_THIS_RUN:
+        return
+
+    setting = util_store_read("beta_translations")
+    css_translations_url = "https://api.deckthemes.com/beta.json" if (setting == "1" or setting == "true") else "https://api.deckthemes.com/stable.json"
+
+    try:
+        async with aiohttp.ClientSession(connector=aiohttp.TCPConnector(verify_ssl=False), timeout=aiohttp.ClientTimeout(total=2)) as session:
+            async with session.get(css_translations_url) as response:
+                if response.status == 200:
+                    with open(css_translations_path, "w", encoding="utf-8") as fp:
+                        fp.write(await response.text())
+
+                        SUCCESSFUL_FETCH_THIS_RUN = True
+                        Log(f"Fetched css translations from server")
+                        initialize_class_mappings()
+                        asyncio.get_running_loop().create_task(loader.reset(silent=True))
+
+    except Exception as ex:
+        Log(f"Failed to fetch css translations from server: {str(ex)}")
+        
+
+async def every(__seconds: float, func, *args, **kwargs):
+    while True:
+        await func(*args, **kwargs)
+        await asyncio.sleep(__seconds)
 
 class FileChangeHandler(FileSystemEventHandler):
     def __init__(self, loader : Loader, loop):
@@ -158,7 +190,7 @@ class Plugin:
         self.server_loaded = False
 
         Log("Initializing css loader...")
-        await initialize_class_mappings()
+        initialize_class_mappings()
         Log(f"Max supported manifest version: {CSS_LOADER_VER}")
         
         create_steam_symlink()
@@ -176,6 +208,8 @@ class Plugin:
         if (ALWAYS_RUN_SERVER or store_or_file_config("server")):
             await self.enable_server(self)
 
+        css_translations_path = os.path.join(get_theme_path(), "css_translations.json")
+        asyncio.get_event_loop().create_task(every(60, fetch_class_mappings, css_translations_path, self.loader))
         await initialize()
 
 if __name__ == '__main__':
